Auton Pushed with most of AlignToTarget Fixed
I had a lot of errors with AlignToTarget because of how it was passed in the constructor. I have only changed how it was the object was passed through it along with the addition of a new ProfiledPIDController because it solved the error of an super constructor initialization problem. The ternary operator of "::" is still an error. Auton is supposed to start on top red tarmac. The robot shoots preloaded cargo and then reverses to the ball behind it. After it shoots that ball, it parks .